added 2 more puzzle sizes
removed the dilate erode morphology as it seems unneccessary with a lower threshold.  Unfortunately it seems that we need need to either create our own puzzles or find a higher resolution source if we want to get to 4000 pieces as above 20x20 the images generated are no longer increasing in resolution and so the segmentation starts falling to the noise.